Release 0.2.0 (#16)

+++ tests/testutil/SourceFileManager.cpp
@@ -6,13 +6,15 @@
 using namespace ftg;
 namespace fs = std::experimental::filesystem;
 
-SourceFileManager::SourceFileManager()
-    : BaseDir(getUniqueFilePath(getTmpDirPath(), "FTGTestSource")) {
+SourceFileManager::SourceFileManager(bool Debug)
+    : Debug(Debug),
+      BaseDir(getUniqueFilePath(getTmpDirPath(), "FTGTestSource")) {
   fs::create_directories(BaseDir);
+  fs::create_directories(getSrcDirPath());
 }
 
 SourceFileManager::~SourceFileManager() {
-  if (fs::exists(BaseDir))
+  if (!Debug && fs::exists(BaseDir))
     fs::remove_all(BaseDir);
 }
 
@@ -26,18 +28,24 @@ bool SourceFileManager::addFile(std::string Path) {
 }
 
 bool SourceFileManager::createFile(std::string Name, std::string Content) {
-  auto Path = fs::path(BaseDir) / Name;
+  auto Path = fs::path(getSrcDirPath()) / Name;
   if (!util::saveFile(Path.c_str(), Content.c_str()))
     return false;
   ManagedFiles.emplace(Name, Path);
   return true;
 }
 
-std::string SourceFileManager::getBaseDirPath() const { return BaseDir; }
-
 std::string SourceFileManager::getFilePath(std::string Name) const {
   auto Iter = ManagedFiles.find(Name);
   if (Iter == ManagedFiles.end())
     return "";
   return Iter->second;
+}
+
+std::string SourceFileManager::getOutDirPath() const {
+  return (fs::path(BaseDir) / fs::path("out")).string();
+}
+
+std::string SourceFileManager::getSrcDirPath() const {
+  return (fs::path(BaseDir) / fs::path("src")).string();
 }
